% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in Fatality Analysis Reporting System data file#'}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{csv type database}
}
\value{
Read in Fatality Analysis Reporting System data file from a csv file
}
\description{
Read in Fatality Analysis Reporting System data file#'
}
\details{
The function will read the data from the path as data frame tibble. If it doesn't exist an error message will be returned.
}
\examples{
WD <- setwd(system.file("extdata", package = "buildingpackagefars"))
fars_read("accident_2013.csv.bz2")
setwd(WD)

}
