% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in multiple Fatality Analysis Reporting System data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{one or multiple years relating to the file names to be read in}
}
\value{
Read in multiple Fatality Analysis Reporting System data files
}
\description{
Read in multiple Fatality Analysis Reporting System data files
}
\details{
The function will read multiple data from the path as data frame tibble. If it doesn't exist an error message will be returned.
}
\examples{
WD <- setwd(system.file("extdata", package = "buildingpackagefars"))
fars_read_years(2013:2015)
setwd(WD)

}
