% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize observations by multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{one or multiple years relating to the file names to be read in and summarize}
}
\value{
Read data and summarize observations by one or more years
}
\description{
Summarize observations by multiple years
}
\details{
The function will read multiple data from the path as data frame tibble nad summerize it by years
}
\examples{
WD <- setwd(system.file("extdata", package = "buildingpackagefars"))
fars_summarize_years(years = c(2013,2014))
setwd(WD)

}
