% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename with a provided year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year to add to the filename}
}
\value{
Create a filename for the accident csv file based on a year.
}
\description{
Create filename with a provided year
}
\details{
Create a filename with a provided year. E.g. if 2015 is provided then it will return "accident_2015_csv.bz2".
}
\examples{
WD <- setwd(system.file("extdata", package = "buildingpackagefars"))
make_filename(2015)
setwd(WD)

}
