% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot State Accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{ID number of a state}

\item{year}{selected year}
}
\value{
Plot State Accidents by year ans state
}
\description{
Plot State Accidents
}
\details{
The function will read data and create a map of accidents for a given state and year.
}
\examples{
library(mapdata)
fars_map_state(12, 2014)
fars_map_state(36, 2014)

\dontrun{
fars_map_state(3, 2014)   # error
}


}
