% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize observations by multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{one or multiple years relating to the file names to be read in and summarize}
}
\value{
Read data and summarize observations by one or more years
}
\description{
Summarize observations by multiple years
}
\details{
The function will read multiple data from the path as data frame tibble nad summerize it by years
}
\examples{
fars_summarize_years(2014:2016)
f13path <- system.file("extdata", "accident_2013.csv.bz2", package = "buildingpackagefars")
f14path <- system.file("extdata", "accident_2014.csv.bz2", package = "buildingpackagefars")
f15path <- system.file("extdata", "accident_2015.csv.bz2", package = "buildingpackagefars")
file.copy(from=c(f13path,f14path,f15path),to=getwd())

}
